% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{.gencoreBin}
\alias{.gencoreBin}
\title{Call the gencore binary with additional arguments.}
\usage{
.gencoreBin(bin = "gencore", args = "", execute = TRUE)
}
\arguments{
\item{bin}{The name of the gencore binary}

\item{args}{A character string containing the arguments that will be passed
to the binary.}

\item{execute}{Logical scalar, whether to execute the command. If FALSE,
return a string with the shell command.}
}
\value{
If \code{execute} is TRUE, returns the console output of running the
  gencore command. If \code{execute} is FALSE, returns the shell command.
}
\description{
Adapted from the gencore package
}
\keyword{internal}
