% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{qcSummary}
\alias{qcSummary}
\title{Summary of Quality Control}
\usage{
qcSummary(json)
}
\arguments{
\item{json}{the output json of function rfastq. [json]}
}
\value{
a data frame.
}
\description{
generate a data frame of the QC summary.
}
\examples{

outputPrefix <- tempfile(tmpdir = tempdir())
se_read1 <- system.file("extdata","Fox3_Std_small.fq.gz",package="Rfastp")
se_json_report <- rfastp(read1 = se_read1, outputFastq = outputPrefix,
   thread = 4)
df_summary <- qcSummary(se_json_report)

}
\author{
Wei Wang
}
