% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{rgencore}
\alias{rgencore}
\title{R wrap of gencore}
\usage{
rgencore(
  inBam = "",
  outBam = "",
  ref = "",
  bed = "",
  umiPrefix = "",
  supportingReads = 1,
  ratioMajorBase = 0.8,
  scoreMajorBase = 6,
  highQual = 30,
  moderateQual = 20,
  lowQual = 15,
  coverageSampling = 10000,
  debug = FALSE,
  quitAfterContig = ""
)
}
\arguments{
\item{inBam}{read1 input file name of sorted bam/sam [string]}

\item{outBam}{file name of output bam/sam [string]}

\item{ref}{Path to reference fasta file. [string]}

\item{bed}{bedfile to specify the capturing region. [string]}

\item{umiPrefix}{the prefix for UMI, if it has. None by default.}

\item{supportingReads}{only output consensus reads/pairs that 
merged by >= `supportingRead` reads/pairs. The valud should be 1~10, and the 
default value is 1.}

\item{ratioMajorBase}{if the ratio of the major base in a cluster is less than 
`ratioMajorBase`, it will be further compared to the reference. The valud should 
be 0.5~1.0, and the default value is 0.8}

\item{scoreMajorBase}{if the score of the major base in a cluster is less than 
`scoreMajorBase`, it will be further compared to the reference. The valud should 
be 1~20, and the default value is 6}

\item{highQual}{the threshold for a quality score to be considered as high qulity.
Default 30 means Q30}

\item{moderateQual}{the threshold for a quality score to be considered as moderate 
qulity. Default 20 means Q20.}

\item{lowQual}{the threshold for a quality score to be considered as low qulity.
Default 15 means Q15.}

\item{coverageSampling}{the sampling ratefor genome scale coverage statistics.
Default 10000 means 1/10000.}

\item{debug}{a logical indicating output some debug information to STDERR.}

\item{quitAfterContig}{stop when `quitAfterContig` contigs are processed. 
Only used for fast debugging. Default 0 means no limitation.}
}
\value{
returns a json object of the report.
}
\description{
Quality control (Cut adapter, low quality trimming, UMI handling, and etc.) of fastq files.
}
\examples{

require(rjson)

# preprare for the input and output files.
# if the output file exists, it will be OVERWRITEN.

inputbamfile <-  system.file("extdata", "ex1.bam", package="Rsamtools")
outputbamfile <- "ex1_rgencore.bam"
reference <- "myreference.fa"

# run rgencore 

rgencore_json_report <- rgencore(inBam=inputbamfile, 
         outBam=outputbamfile,
         ref=reference
)
}
\author{
Thomas Carroll, Wei Wang
}
