% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{catfastq}
\alias{catfastq}
\title{concatenate file in R}
\usage{
catfastq(output, inputFiles, append = FALSE, paired = FALSE)
}
\arguments{
\item{output}{output file name [string]}

\item{inputFiles}{a list of input file names [list]}

\item{append}{a logical indicating append the files to a file already
exists.}

\item{paired}{a logical indicating split the input files into two halves.
the first half merged into read1, the second half merged into read2.}
}
\value{
no returns.
}
\description{
concatenate multiple files into a big file.
}
\examples{

pe001_read1 <- system.file("extdata","splited_001_R1.fastq.gz",package="Rfastp")
pe002_read1 <- system.file("extdata","splited_002_R1.fastq.gz",package="Rfastp")
pe003_read1 <- system.file("extdata","splited_003_R1.fastq.gz",package="Rfastp")
pe004_read1 <- system.file("extdata","splited_004_R1.fastq.gz",package="Rfastp")

pe001_read2 <- system.file("extdata","splited_001_R2.fastq.gz",package="Rfastp")
pe002_read2 <- system.file("extdata","splited_002_R2.fastq.gz",package="Rfastp")
pe003_read2 <- system.file("extdata","splited_003_R2.fastq.gz",package="Rfastp")
pe004_read2 <- system.file("extdata","splited_004_R2.fastq.gz",package="Rfastp")

allR1 <- list(pe001_read1, pe002_read1, pe003_read1, pe004_read1)
allR2 <- list(pe001_read2, pe002_read2, pe003_read2, pe004_read2)

allreads <- append(allR1, allR2)

# a normal single-end concatenation.

catfastq(output = "merged1_R1.fastq.gz", inputFiles = allR1)

# a normal paired-end concatenation.

catfastq(output = "merged_paired", inputFiles = allreads, paired=TRUE)

# append to exists files (paired-end)

catfastq(output="append_paired", inputFiles=allreads, append=TRUE, paired=TRUE)

}
\author{
Thomas Carroll, Wei Wang
}
